Optimize single page results: avoid performing a COUNT query
When a last page is requested (meaing there are less results then the page size),
FeatureCollection.number_matched can avoid a COUNT query by reading number_result.

As a consequence, the whole queryset results are read at an earlier
stage. Hence, the decorate_queryset() logic needs to be called earlier
too. This is why the DB-logic is now wrapped in classmethods.