Implemented basic support for rendering complex field values.
* Added `XsdComplexType` definition, which `FeatureType` may expose.
* Remoddeled DescribeFeatureType to work directly with <xsd:ComplexType> elements.
* Cleanup in decorate_queryset() to allow extensions later.

This is preparation work for complex feature types, such as embedded
fields of a different subtype. Instead of rendering a list of 'xs:type'
elements, a series of complextypes can now be defined.

The logic for defining this easily in FeatureType is not designed yet.
Currently, the unit tests override `get_field_type()` for this.