Fix FES filtering for ArrayField members
A default "field = value" will not work unless the value is an array
type itself. Since these fields are rendered as multiple XML nodes
(similar to M2M/reverse FK), it makes more sense to apply the query on
single array members instead. This is achieved using the PostgreSQL ANY() operator.
So "arrayfield < value" is queried as "value > ANY(arrayfield)".