Optimize buffer writing: 15% faster
The buffer size is now 8192 bytes, and the BaseBuffer.size is gone,
since we can use BaseBuffer.data.tell instead.

The following microbenchmark...

    >>> from buffer import BytesBuffer
    >>> buf = BytesBuffer()
    >>> def writebuf(data, n):
    ...     for i in range(n):
    ...         buf.write(data)
    ...         if buf.is_full():
    ...             buf.flush()
    ...     buf.flush()
    ...
    >>> %timeit writebuf(b"foobar", 20000)

shows a 16% speedup. Before:

    12.8 ms ± 121 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

After:

    10.7 ms ± 15.2 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

The improvement is almost the same for a version that uses StringBuffer.
Before:

    12.9 ms ± 150 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

After:

    10.6 ms ± 117 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

A DSO-API with "fietspaaltjes" loaded can now serve the full dataset at
4.1MB/s (from localhost), vs. previously 3.6MB/s, in GML 3.2 format.