Change exception output during streaming results, changed exception class signature!
The 'locator' argument turns out to be optional, hence it's no longer
the first parameter. This also makes the exception class signature in
line with regular Python exceptions.

The WSF XSD spec has a hidden gem: an <wfs:truncatedResponse> element
can be rendered to inform the client an error happened during rendering.