Implemented PROPERTYNAME support
While this principle is simple at a glance, proper support
redefined how output is rendered. The new FeatureProjection class defines
which fields and child elements will be rendered.

Instead of inspecting the FeatureType/XsdComplexType data,
it looks at the FeatureProjection instead. This information is
calculated once, so rendering 5000 records should remain fast.

This same information is used to supply the queryset .only()
and .prefetch_related() with less fields to query so these queries are
as optimal as possible too.